pyproject.toml
[tool.poetry]
name = "gobc"
version = "0.1.0"
description = "Summarizer for GOBC"
authors = ["David Godinez <davidgodinez70@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
openai = "^0.27.8"
python-docx = "^0.8.11"
tk = "^0.1.0"
Pillow = "^10.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
